# Format/size detection performance improvements