--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -11,7 +11,6 @@ on:
 jobs:
   tests:
     runs-on: ubuntu-latest
-    needs: config
     timeout-minutes: 5
     steps:
       - uses: actions/checkout@v4