--- a/tests/test_data/corrupted.png
+++ b/tests/test_data/corrupted.png
@@ -1,0 +1,3 @@
+‰PNG
+
+corrupted data