--- a/tests/test_data/test.md
+++ b/tests/test_data/test.md
@@ -1,0 +1,33 @@
+![Remote 512x512.png](https://placehold.co/512x512.png)
+
+![Remote 512x512.jpg](https://placehold.co/512x512.jpg)
+
+![Remote 512x512.gif](https://placehold.co/512x512.gif)
+
+![100x100.png](100x100.png)
+
+![256x256.ico](256x256.ico)
+
+![256x256.png](256x256.png)
+
+![512x512.avif](512x512.avif)
+
+![512x512.bmp](512x512.bmp)
+
+![512x512.gif](512x512.gif)
+
+![512x512.heic](512x512.heic)
+
+![512x512.JPEG](512x512.JPEG)
+
+![512x512.jpg](512x512.jpg)
+
+![512x512.png](512x512.png)
+
+![512x512.svg](512x512.svg)
+
+![512x512.webp](512x512.webp)
+
+![512x512.xpm](512x512.xpm)
+
+![512x512_actually_jpeg.png](512x512_actually_jpeg.png)