… .gitignore
… tests/test_data/100x100.png
… tests/test_data/256x256.ico
… tests/test_data/256x256.png
… tests/test_data/512x512.JPEG
… tests/test_data/512x512.avif
… tests/test_data/512x512.bmp
… tests/test_data/512x512.gif
… tests/test_data/512x512.heic
… tests/test_data/512x512.jpg
… tests/test_data/512x512.png
… tests/test_data/512x512.svg
… tests/test_data/512x512.webp
… tests/test_data/512x512.xpm
… tests/test_data/512x512_actually_jpeg.png
… tests/test_data/test.txt