Adds fast, lua-based format detection and dimension reading for common image formats, falling back to ImageMagick when needed. This avoids spawning external processes for basic operations, making things noticeably snappier, especially when using image.nvim's API from other plugins (my use-case).

- Extended built-in format detection using magic numbers instead of relying on ImageMagick
- Direct header parsing for image dimensions (PNG, JPEG, WebP, GIF, BMP, ICO, HEIC, AVIF)
- Added test suite with sample images and edge cases
- Set up basic GitHub Actions CI for running tests

The lua implementations are significantly faster but slightly less forgiving with malformed files compared to ImageMagick. We fall back to the existing methods when lua-based parsing fails, so everything should be backward compatible. More test images might be beneficial, especially for heic/avif formats.

Benchmarking the time to run `from_image`:

- With `magick_rock`:
  - For a 2880 x 1800 PNG file:
    - Before: `283.36ms`
    - After: `0.50ms`
  - For a 208 x 209 PNG file:
    - Before: `4.96ms`
    - After: `0.36ms`
- With `magick_cli`:
  - For a 2880 x 1800 PNG file:
    - Before: `294.32ms`
    - After: `0.48ms`
  - For a 208 x 209 PNG file:
    - Before: `32.82ms`
    - After: `0.16ms`

Similar speedups for other formats. 

Note: most of the logic in `dimensions.lua` was generated with Anthropic's Claude.